Massive refactoring step affecting color range and palette definitions
I apparently decided back in the day that ranges/palettes should be
stored in unsorted lists of wrapper objects associating ids and names to
them. Now, I have decided that we only really need ids for custom items,
and now they should double as names in the UI, *except* built-in items,
which are treated in a rather special manner by some code.

In order to disentangle this mess, a lot of code was moved around.
Generally speaking, this should be an improvement (especially the UI
setup code in the MainWindow class). However, one should always be
prepared to fix regressions after massive changes like this.

At least everything appears to be working fine on my end.