Make UI definitions work with Qt 4.6
Although the QIcon::fromTheme() was apparently introduced in 4.6, the
relevant UI definition syntax for setting theme icons in the .ui files
is more recent (possibly 4.8).

This fix works by moving the theme icon specifications out of the .ui
files and taking care of them in the C++ in the window class
constructors.