Fix issue with the process not exiting after closing the main window
In loonycyborg's words:

  <loonycyborg> [...] the only thing I know is that if I replace that QFileDialog() call with hardcoded string it exits on window close

Turns out we're not supposed to call any UI functions before running the
event loop, otherwise things get *weird* to say the least. Apparently on
other platforms it's even possible for the app to crash.

This does not solve a new issue wherein the main window does not regain
focus on Plasma 5.13 (and probably others) after closing the file
dialog, even though the file dialog is supposed to be owned by the main
window per MainWindow::do_open().