Make color icons a bit more device pixel density-independent
This makes color icons look correct on high and low DPI screens alike.
However, we currently only render the icon when it's assigned to items
and not when the window is moved to a different screen, so we still
have an issue where an icon will look completely wrong if it's moved
to a screen with a different pixel ratio than what the original used
at construction time.