paletteitem: Additional overloads of createColorIcon()
This allows callers to pass a ColorRange directly. We use the rep()
value from it automatically.

Note that this also adds a QRgb overload of the function. This is
important as otherwise compilers will run into issues due to ambiguity
with the ColorRange and QColor constructors, since both can take
exactly one QRgb value as their first and only parameter.