settingsdialog: Add options to use Wesnoth built-ins as templates
This allows the creation of new colour ranges and palettes that are
copies of the built-ins provided by Wesnoth much faster than before,
by not needing to manually copy their colour values from WML and use
the From List dialog. The idea here is to allow easier experimentation
with variants of mainline objects.