Add modified .kra and .ora plugins from KImageFormats 6.1
These are altered to use QuaZip instead of KArchive. Amazingly, the
changed needed to do so are pretty minimal and only affect the read()
method of each format handler class.

The reason for changing the backend used for reading zip files is that
I ran into issue after issue trying to use the upstream versions of
KArchive and KImageFormats with our toolchain on macOS. The fact is
that while the KDE Frameworks' code is readily available for use in
3rd-party projects, the documentation for doing so (especially on a
non-integrated environment such as Windows or macOS, where you can't
rely on the libraries being already available system-wide) is lacking
to the point I can't find it. I love KDE but, please.

(Part of #7.)