main: Ensure event MIME data is not a null pointer
The KDE Wayland porting guide mentions this:

  QClipboard::mimeData() can return nullptr on Wayland, leading to
  crashes in code assuming that the mimeData pointer is always valid.
  Always check the validity of the pointer returned by
  QClipboard::mimeData().

<https://community.kde.org/Guidelines_and_HOWTOs/Wayland_Porting_Notes#QClipboard::mimeData()>

We don't interact with QClipboard::mimeData() directly, but we do use
QDropEvent::mimeData(). I don't entirely trust the documentation above
to be exhaustive, and I'm not entirely sure how to intentionally cause
a scenario where it would potentially make sense for the value to be
null, so let's just err on the safe side here.