main: Make drag/pan event implementation less dumb
Make drag/pan not use deprecated pos(), fix composite preview

We weren't tracking the event position relative to the composite
preview widget for mousePressEvent(), resulting in dragging from the
composite preview widget only working if at some point its rectangle
overlaps the split preview widgets, which would only happen after
messing with the view mode selection in the UI.

This also makes it so we don't actually check if the drag event start
position is contained in widgets that aren't currently visible if the
current view mode does not allow for that.

Finally, we no longer use the deprecated pos() method.