readme: Update instructions for macOS Sequoia
macOS Sequoia removes the Finder-only method of bypassing Gatekeeper.
This patch replaces the instructions with the official Sequoia-friendly
method. I shall assume users on older versions already know how to deal
with Gatekeeper, just so I don't need to write a whole essay on this --
the differences should be pretty obvious anyway for Monterey and older
users.