Add dark mode support on Windows using the Fusion Qt style engine
Unfortunately our options here are pretty limited. The only native
engine that supports both light and dark mode properly on Windows is
the Windows 11 style engine, which continues to suck (maybe our usage
of groupboxes is the root cause). Thus we are forced to use Fusion for
dark mode support instead.

Since we don't want to use Fusion by default, we are also forced to
disable the automatic color scheme option on Windows and pretend it's
light mode by default instead. Ideally we would detect the color scheme
change, but the QStyleHints API is insufficient for this purpose; we'd
need to closely inspect the palette on QEvent::PaletteChange and I
really don't feel like doing that right now.

Meanwhile, on Linux, we still can't use dark mode even if Fusion is the
active style engine. Once again, it may be necessary to add
functionality to select our own palette like some KDE applications do
(such as Konsole, Kdenlive, and Kate).