paletteitem: Use global palette to source color icon borders
The widget's palette is completely incorrect on Windows using Fusion.
Even if we source ButtonText or WindowText, we get a dark color instead
of something readable against a dark background. I'm not sure how this
is even possible, but for the sake of getting this functionality out
we'll just have to use the global palette instead of the widgets'
individual palettes.

(Tested on Qt 6.9.3.)