main: Use palettes to set the preview area background color
By not using stylesheets, we don't accidentally change the background
color of the scrollbars of the QScrollArea, which caused them to turn
bright in dark mode whenever a non-default background color was set.

(It *is* possible to workaround the issue in question while still using
stylesheets, but it either results in the scrollbars being rendered in
a non-styled fashion, or additional inconsistencies. It's best to just
avoid throwing a whole stylesheet at this issue.)